Updates and Changes

+++ code/game/objects/effects/aliens.dm
@@ -4,6 +4,7 @@
  #define ACID_STRONG     2
  #define ACID_MODERATE   1.5
  #define ACID_WEAK       1
+
 /obj/effect/acid
 	name = "acid"
 	desc = "Burbling corrosive stuff. Probably a bad idea to roll around in it."
@@ -15,31 +16,41 @@
 	anchored = 1
 
 	var/atom/target
-	var/ticks = 0
 	var/acid_strength = ACID_WEAK
 	var/melt_time = 10 SECONDS
+	var/last_melt = 0
 
 /obj/effect/acid/New(loc, supplied_target)
 	..(loc)
 	target = supplied_target
 	melt_time = melt_time / acid_strength
+	START_PROCESSING(SSprocessing, src)
 
-	melt_stuff()
+/obj/effect/acid/Destroy()
+	STOP_PROCESSING(SSprocessing, src)
+	target = null
+	. = ..()
 
-/obj/effect/acid/proc/melt_stuff()
-	spawn(1)
-		var/melted = target.acid_melt(melt_time)
-		sleep(melt_time)
-		if(melted)
+/obj/effect/acid/Process()
+	if(QDELETED(target))
+		qdel(src)
+	else if(world.time > last_melt + melt_time)
+		var/done_melt = target.acid_melt()
+		last_melt = world.time
+		if(done_melt)
 			qdel(src)
-		else
-			melt_stuff()
 
-/atom/proc/acid_melt(var/melt_time)
-	visible_message("<span class='alium'>Acid hits \the [src] with a sizzle!</span>")
-	for(var/i = 0, i < 4, i++)
-		sleep(melt_time)
-		visible_message("<span class='alium'>The acid melts \the [src]!</span>")
-	visible_message("<span class='alium'>The acid melts \the [src] away into nothing!</span>")
-	qdel(src)
-	return TRUE
+/atom/var/acid_melted = 0
+
+/atom/proc/acid_melt()
+	. = FALSE
+	switch(acid_melted)
+		if(0)
+			visible_message("<span class='alium'>Acid hits \the [src] with a sizzle!</span>")
+		if(1 to 3)
+			visible_message("<span class='alium'>The acid melts \the [src]!</span>")
+		if(4)
+			visible_message("<span class='alium'>The acid melts \the [src] away into nothing!</span>")
+			. = TRUE
+			qdel(src)
+	acid_melted++
